+++ src/main/java/vectorwing/farmersdelight/common/registry/ModSounds.java
@@ -56,7 +56,13 @@ public class ModSounds
 	// Tomato Bush
 	public static final RegistryObject<SoundEvent> ITEM_TOMATO_PICK_FROM_BUSH = SOUNDS.register("block.tomato_bush.pick_tomatoes",
 			() -> SoundEvent.createVariableRangeEvent(new ResourceLocation(FarmersDelight.MODID, "block.tomato_bush.pick_tomatoes")));
-	
+
+	// Food
+	public static final RegistryObject<SoundEvent> BLOCK_TAKE_PORTION = SOUNDS.register("block.food.take_portion",
+			() -> SoundEvent.createVariableRangeEvent(new ResourceLocation(FarmersDelight.MODID, "block.food.take_portion")));
+	public static final RegistryObject<SoundEvent> BLOCK_PASTRY_SLICE = SOUNDS.register("block.pastry.slice",
+			() -> SoundEvent.createVariableRangeEvent(new ResourceLocation(FarmersDelight.MODID, "block.pastry.slice")));
+
 	public static final RegistryObject<SoundEvent> ENTITY_ROTTEN_TOMATO_THROW = SOUNDS.register("entity.rotten_tomato.throw",
 			() -> SoundEvent.createVariableRangeEvent(new ResourceLocation(FarmersDelight.MODID, "entity.rotten_tomato.throw")));
 	public static final RegistryObject<SoundEvent> ENTITY_ROTTEN_TOMATO_HIT = SOUNDS.register("entity.rotten_tomato.hit",